lag database & client package i infrastructure for dokument
