+++ client/src/main/kotlin/no/nav/su/se/bakover/client/journalfør/skatt/påsak/JournalførSkattedokumentPåSakFakeClient.kt
@@ -4,7 +4,7 @@ import arrow.core.Either
 import arrow.core.right
 import no.nav.su.se.bakover.common.domain.client.ClientError
 import no.nav.su.se.bakover.common.journal.JournalpostId
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalpostIdGeneratorForFakes
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalpostIdGeneratorForFakes
 import no.nav.su.se.bakover.domain.journalpost.JournalførSkattedokumentPåSakCommand
 import no.nav.su.se.bakover.domain.skatt.JournalførSkattedokumentPåSakClient
