+++ client/src/main/kotlin/no/nav/su/se/bakover/client/journalfør/skatt/påsak/JournalførSkattedokumentPåSakHttpClient.kt
@@ -3,14 +3,13 @@ package no.nav.su.se.bakover.client.journalfør.skatt.påsak
 import arrow.core.Either
 import no.nav.su.se.bakover.common.domain.client.ClientError
 import no.nav.su.se.bakover.common.journal.JournalpostId
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.Fagsak
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalPostType
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalførHttpClient
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalførJsonRequest
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalførendeEnhet
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalpostDokument
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.tilBehandlingstema
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.tilBruker
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.Fagsak
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalPostType
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalførJsonRequest
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalførendeEnhet
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalpostDokument
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.tilBehandlingstema
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.tilBruker
 import no.nav.su.se.bakover.domain.journalpost.JournalførSkattedokumentPåSakCommand
 import no.nav.su.se.bakover.domain.skatt.JournalførSkattedokumentPåSakClient
 
@@ -18,7 +17,7 @@ import no.nav.su.se.bakover.domain.skatt.JournalførSkattedokumentPåSakClient
  * TODO jah: Flytt til skatt modul når vi har en skatt modul.
  */
 internal class JournalførSkattedokumentPåSakHttpClient(
-    private val client: JournalførHttpClient,
+    private val client: no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalførHttpClient,
 ) : JournalførSkattedokumentPåSakClient {
     override fun journalførSkattedokument(command: JournalførSkattedokumentPåSakCommand): Either<ClientError, JournalpostId> {
         val dokument = command.dokument
