+++ client/src/test/kotlin/no/nav/su/se/bakover/client/journalfør/JournalpostCommandTest.kt
@@ -1,7 +1,7 @@
 package no.nav.su.se.bakover.client.journalfør
 
 import io.kotest.matchers.shouldBe
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.DokumentVariant
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.DokumentVariant
 import org.junit.jupiter.api.Test
 
 internal class JournalpostCommandTest {
