+++ database/src/main/kotlin/no/nav/su/se/bakover/database/revurdering/AvsluttetRevurderingDatabaseJson.kt
@@ -1,7 +1,7 @@
 package no.nav.su.se.bakover.database.revurdering
 
 import no.nav.su.se.bakover.common.tid.Tidspunkt
-import no.nav.su.se.bakover.dokument.infrastructure.BrevvalgDbJson
+import no.nav.su.se.bakover.dokument.infrastructure.database.BrevvalgDbJson
 
 data class AvsluttetRevurderingDatabaseJson(
     val begrunnelse: String,
