+++ dokument/infrastructure/src/main/kotlin/no/nav/su/se/bakover/dokument/infrastructure/database/Dokumentkomponenter.kt
@@ -1,4 +1,4 @@
-package no.nav.su.se.bakover.dokument.infrastructure
+package no.nav.su.se.bakover.dokument.infrastructure.database
 
 import no.nav.su.se.bakover.dokument.application.DokumentServices
 
@@ -6,6 +6,6 @@ import no.nav.su.se.bakover.dokument.application.DokumentServices
  * Et forsøk på modularisering av [no.nav.su.se.bakover.web.services.Services] og [no.nav.su.se.bakover.domain.DatabaseRepos] der de forskjellige modulene er ansvarlige for å wire opp sine komponenter.
  */
 class Dokumentkomponenter(
-    val repos: DokumentRepos,
+    val repos: no.nav.su.se.bakover.dokument.infrastructure.database.DokumentRepos,
     val services: DokumentServices,
 )
