+++ dokument/infrastructure/src/main/kotlin/no/nav/su/se/bakover/dokument/infrastructure/database/journalføring/JournalpostIdGeneratorForFakes.kt
@@ -1,4 +1,4 @@
-package no.nav.su.se.bakover.dokument.infrastructure.journalføring
+package no.nav.su.se.bakover.dokument.infrastructure.database.journalføring
 
 import no.nav.su.se.bakover.common.journal.JournalpostId
 import java.util.concurrent.atomic.AtomicLong
