+++ dokument/infrastructure/src/main/kotlin/no/nav/su/se/bakover/dokument/infrastructure/database/journalføring/brev/JournalførBrevHttpClient.kt
@@ -1,21 +1,20 @@
-package no.nav.su.se.bakover.dokument.infrastructure.journalføring.brev
+package no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.brev
 
 import arrow.core.Either
 import dokument.domain.journalføring.brev.JournalførBrevClient
 import dokument.domain.journalføring.brev.JournalførBrevCommand
 import no.nav.su.se.bakover.common.domain.client.ClientError
 import no.nav.su.se.bakover.common.journal.JournalpostId
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.AvsenderMottaker
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.Fagsak
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalPostType
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalførHttpClient
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalførJsonRequest
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalførendeEnhet
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalpostDokument
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.tilBehandlingstema
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.tilBruker
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.AvsenderMottaker
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.Fagsak
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalPostType
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalførJsonRequest
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalførendeEnhet
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalpostDokument
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.tilBehandlingstema
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.tilBruker
 
-internal class JournalførBrevHttpClient(private val client: JournalførHttpClient) : JournalførBrevClient {
+internal class JournalførBrevHttpClient(private val client: no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalførHttpClient) : JournalførBrevClient {
     override fun journalførBrev(command: JournalførBrevCommand): Either<ClientError, JournalpostId> {
         val dokument = command.dokument
         return client.opprettJournalpost(
@@ -40,6 +39,6 @@ internal class JournalførBrevHttpClient(private val client: JournalførHttpClie
     }
 }
 
-fun createJournalførBrevHttpClient(client: JournalførHttpClient): JournalførBrevClient {
+fun createJournalførBrevHttpClient(client: no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalførHttpClient): JournalførBrevClient {
     return JournalførBrevHttpClient(client)
 }
