+++ dokument/infrastructure/src/main/kotlin/no/nav/su/se/bakover/dokument/infrastructure/database/journalføring/søknad/JournalførSøknadFakeClient.kt
@@ -1,4 +1,4 @@
-package no.nav.su.se.bakover.dokument.infrastructure.journalføring.søknad
+package no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.søknad
 
 import arrow.core.Either
 import arrow.core.right
@@ -6,7 +6,7 @@ import dokument.domain.journalføring.søknad.JournalførSøknadClient
 import dokument.domain.journalføring.søknad.JournalførSøknadCommand
 import no.nav.su.se.bakover.common.domain.client.ClientError
 import no.nav.su.se.bakover.common.journal.JournalpostId
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalpostIdGeneratorForFakes
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalpostIdGeneratorForFakes
 
 /**
  * Åpen for å kunne aksesseres av tester.
