+++ dokument/infrastructure/src/main/kotlin/no/nav/su/se/bakover/dokument/infrastructure/database/journalføring/søknad/JournalførSøknadHttpClient.kt
@@ -1,4 +1,4 @@
-package no.nav.su.se.bakover.dokument.infrastructure.journalføring.søknad
+package no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.søknad
 
 import arrow.core.Either
 import dokument.domain.journalføring.søknad.JournalførSøknadClient
@@ -6,21 +6,20 @@ import dokument.domain.journalføring.søknad.JournalførSøknadCommand
 import no.nav.su.se.bakover.common.domain.client.ClientError
 import no.nav.su.se.bakover.common.domain.sak.Sakstype
 import no.nav.su.se.bakover.common.journal.JournalpostId
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.AvsenderMottaker
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.Fagsak
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalPostType
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalførHttpClient
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalførJsonRequest
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalførendeEnhet
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.JournalpostDokument
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.Kanal
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.tilBehandlingstema
-import no.nav.su.se.bakover.dokument.infrastructure.journalføring.tilBruker
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.AvsenderMottaker
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.Fagsak
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalPostType
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalførJsonRequest
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalførendeEnhet
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalpostDokument
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.Kanal
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.tilBehandlingstema
+import no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.tilBruker
 
 /**
  * TODO jah: Denne bør egentlig bo i :søknad-modulen som ikke eksisterer på dette tidspunktet.
  */
-internal class JournalførSøknadHttpClient(private val client: JournalførHttpClient) : JournalførSøknadClient {
+internal class JournalførSøknadHttpClient(private val client: no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalførHttpClient) : JournalførSøknadClient {
     override fun journalførSøknad(command: JournalførSøknadCommand): Either<ClientError, JournalpostId> {
         return client.opprettJournalpost(
             JournalførJsonRequest(
@@ -50,6 +49,6 @@ internal class JournalførSøknadHttpClient(private val client: JournalførHttpC
     }
 }
 
-fun createJournalførSøknadHttpClient(client: JournalførHttpClient): JournalførSøknadClient {
+fun createJournalførSøknadHttpClient(client: no.nav.su.se.bakover.dokument.infrastructure.database.journalføring.JournalførHttpClient): JournalførSøknadClient {
     return JournalførSøknadHttpClient(client)
 }
