+++ dokument/infrastructure/src/test/kotlin/no/nav/su/se/bakover/dokument/infrastructure/database/distribuering/DokDistFordelingClientTest.kt
@@ -1,4 +1,4 @@
-package no.nav.su.se.bakover.dokument.infrastructure.distribuering
+package no.nav.su.se.bakover.dokument.infrastructure.database.distribuering
 
 import arrow.core.right
 import com.github.tomakehurst.wiremock.client.MappingBuilder
